moved print search results into domPrinter module
